requirements.txt
# requirements.txt
bcrypt==4.1.3
cffi==1.16.0
cryptography==42.0.8
et-xmlfile==1.1.0
numpy==2.0.0
openpyxl==3.1.5
pandas==2.2.2
pip==24.1.1
psycopg2==2.9.9
pycparser==2.22
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
setuptools==63.2.0
six==1.16.0
sshtunnel==0.4.0
tzdata==2024.1